+++ src/pages/shift/Shift.js
@@ -120,7 +120,6 @@ export default {
                         url: self_url,
                         data: this.addParams
                     }).then(res => {
-                        console.log(res);
                         res = res.data;
                         if (res.status === 'SUCCESS') {
                             if (res.content.status === true) {
@@ -172,26 +171,37 @@ export default {
         deleteHandle: function(id) {
             let delArr = [];
             delArr.push(id);
-            this.$ajax({
-                method: 'delete',
-                url: '/shift',
-                data: delArr
-            }).then(res => {
-                res = res.data;
-                console.log(res);
-                if (res.status === 'SUCCESS') {
-                    this.getApi(this.reqData);
-                    this.$message({showClose: true, message: res.msg, type: 'success'})
-                } else {
-                    this.$message({showClose: true, message: res.msg, type: 'warning'})
-                }
-            }).catch(err => {
-                this.$message({
-                    showClose: true,
-                    message: '获取数据失败！',
-                    type: 'warning'
+            this.$confirm('您确定要删除该条数据吗?', '删除', {
+                confirmButtonText: '确定',
+                cancelButtonText: '取消',
+                type: 'warning'
+            }).then(() => {
+                this.$ajax({
+                    method: 'delete',
+                    url: '/shift',
+                    data: delArr
+                }).then(res => {
+                    res = res.data;
+                    console.log(res);
+                    if (res.status === 'SUCCESS') {
+                        this.getApi(this.reqData);
+                        this.$message({showClose: true, message: res.msg, type: 'success'})
+                    } else {
+                        this.$message({showClose: true, message: res.msg, type: 'warning'})
+                    }
+                }).catch(err => {
+                    this.$message({
+                        showClose: true,
+                        message: '获取数据失败！',
+                        type: 'warning'
+                    })
                 })
-            })
+            }).catch(() => {
+                this.$message({
+                    type: 'info',
+                    message: '已取消删除'
+                });
+            });
         },
         showAlert: function() {
             this.showTitle = '新增班次';
